Changed smoking and alcohol proportions
I went into the datafeedback code and changed alcohol to accept both 1 and 0 as an answer. As well I had to change the logic of the alcohol and smoking functions to create a dataframe which selects both 1 and 0 as an answer. I also changed the Codebook.txt to reflect this change in the proportion.

I also changed the working directory for all 3 to be the WD that would be expected if a person moved the folder onto the desktop and ran the code.