Checkbox fix to feedback and clean code 2
I added code to feedback and clean code 2 to make them also display correct values for %kin, which is the  only other checkbox we are doing proportions of. As well, turned the proportion maker in feedback into its own bespoke function, so we can re-use it for other checkboxes if we need. Finally the location of the checker function in clean code 2 was moved up as it is used by both kin and health problems.