Feedback more accurately calcs network size
Changed the feedback code to more accurately calculate the network size. Unfortunately I couldn't use the same method as the clean-code2 as that method has batching built in, while my version is for-looping. The problem was that two separate persons identified in the original network were getting eliminated even thought the respondent indicated they were separate people. Now the code takes this into consideration, while still eliminating box names that are the same as the original names.