Cleaning examples and naming output variable
Cleaned up the example in egodata to better match the description and remove these example variables after the example to not cause any confusion. Also changed the initial notes to indicate the name of the variable we are outputting, for better reference in later code if user does not know what ego_data.rda or alter_data.rda contain.