Fixing package attachment, adding example functions
Fixed standard setup for all 5 data files to automatically install packages if they are not present, including code to access the currently archived EgoNet package. As well added an additional set of code which reworks cleandata2 in the same style as alterdata and egodata, alllowing for easier modification and better tutorialization.